The issue is reproduced because we believed that if an ListViewItem does not have a ListView, then it is inaccessible to the user and has no interaction with the Accessibility tool (Inspect, Narrator).
Unfortunately, after deleting an ListViewItem, the Accessibility tool still tries to get information about it, although the ListViewItem no longer has a ListView. Therefore, the logic has been updated to work with an ListViewItem that does not have a ListView.

Also, after the focused ListViewItem is removed, it remains cached in the "FocusedItem" property. It also forces the Accessibility tool to try to retrieve the data of the ListViewItem that is not displayed. As a fix, a logic was added to reset the "Focused" flag from the ListViewItem.

# Conflicts:
#	src/System.Windows.Forms/src/System/Windows/Forms/ListViewItem.ListViewItemBaseAccessibleObject.cs
#	src/System.Windows.Forms/src/System/Windows/Forms/ListViewItem.cs
#	src/System.Windows.Forms/tests/UnitTests/System/Windows/Forms/ListViewTests.cs